+++ NEWS.md
@@ -2,7 +2,9 @@
 * Parameters added for downsampling/thinning the dataset prior to plotting with manhattan
 * Locuszoom plot updated to include variants with missing R2 values
 * vline_color bug fix
-* improved get_lead_snp functionality
+* Improved get_lead_snp functionality
+* The get_sign_and_sugg_loci function added for extracting genome-wide significant and suggestive loci
+* The manhattanExtra function added for highlighting genome-wide significant and suggestive loci
 
 # 2.0.1
 * REF and ALT error message fixed when there are p-values below .Machine$double.xmin in the input dataset